// backend/src/student/dto/student-document.dto.ts
export class StudentDocumentDto {
  photo: string;
  passport: string;
  pg: string;
  pg_marksheet: string;
  ug: string;
  ug_marksheet: string;
  higher_sec: string;
  tenth: string;
  lor: string;
  moi: string;
  cv: string;
  exp_cert: string;
  sop: string;
  other_diploma_cert: string;
  others: string;
}